% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MexicanHH_foodConsumption}
\alias{MexicanHH_foodConsumption}
\title{National Survey of Household Income and Expenditures (ENIGH)}
\format{
## `MexicanHH_foodConsumption`
A data frame with 8,777 rows and 17 columns:
\describe{
 \item{s1}{Share of tortilla consumption by household.}
 \item{s2}{Share of cereal consumption by household.}
 \item{s3}{Share of meat consumption by household.}
 \item{s4}{Share of dairy consumption by household.}
 \item{s5}{Share of fruits and vegetables consumption by household.}
 \item{s6}{Share of other consumption by household.}
 \item{lnp1}{Logged price of tortilla.}
 \item{lnp2}{Logged price of cereal.}
 \item{lnp3}{Logged price of meat.}
 \item{lnp4}{Logged price of dairy.}
 \item{lnp5}{Logged price of fruits and vegetables.}
 \item{lnp6}{Logged price of other.}
 \item{lnw}{Logged food expenditure by household .}
 \item{age}{Logged age of head of household.}
 \item{size}{Inverse (1/x) of size of household.}
 \item{sex}{Sex (female == 1) of head of household.}
 \item{educ}{Educational attaintment of head of household.}
}


' @source <https://en.www.inegi.org.mx/programas/enigh/nc/2022/>
}
\usage{
MexicanHH_foodConsumption
}
\description{
MexicanHH_foodConsumption is a 10 percent sample of the data described in Beckman et al. (2024)'s Land Competition and Welfare Effects from Mexico's Proposal to Ban Genetically Modified Corn. An overview of the data construction and assessment is further discussed in the study.
}
\keyword{datasets}
